# packages/admin-client/openapi/admin.openapi.json
{
  "components": {
    "schemas": {
...
      "CallLifecycleResponse": {
        "additionalProperties": false,
        "properties": {
          "call_session_id": {
            "format": "uuid",
            "title": "Call Session Id",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/CallLifecycleStatus"
          }
        },
        "required": [
          "call_session_id",
          "status",
          "started_at",
          "ended_at",
          "failure_reason"
        ],
        "title": "CallLifecycleResponse",
        "type": "object"
      },
      "CallLifecycleStatus": {
        "enum": [
          "created",
          "active",
          "completed",
          "failed"
        ],
        "title": "CallLifecycleStatus",
        "type": "string"
      },
...
      "CreateTestVoiceSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id"
        ],
        "title": "CreateTestVoiceSessionRequest",
        "type": "object"
      },
      "CreateTestVoiceSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "call_session_id": {
            "format": "uuid",
            "title": "Call Session Id",
            "type": "string"
          },
          "livekit_url": {
            "title": "Livekit Url",
            "type": "string"
          },
          "participant_identity": {
            "title": "Participant Identity",
            "type": "string"
          },
          "participant_token": {
            "title": "Participant Token",
            "type": "string"
          },
          "room_name": {
            "title": "Room Name",
            "type": "string"
          }
        },
        "required": [
          "call_session_id",
          "room_name",
          "livekit_url",
          "participant_identity",
          "participant_token"
        ],
        "title": "CreateTestVoiceSessionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Agent Platform Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/v1/voice/test-sessions": {
      "post": {
        "operationId": "create_test_voice_session_admin_v1_voice_test_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestVoiceSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTestVoiceSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Create Test Voice Session",
        "tags": [
          "admin:voice"
        ]
      }
    },
    "/admin/v1/voice/test-sessions/{call_id}": {
      "get": {
        "operationId": "get_test_voice_session_admin_v1_voice_test_sessions__call_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Call Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallLifecycleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Get Test Voice Session",
        "tags": [
          "admin:voice"
        ]
      }
    }
  }
}